Refactor CountryDrawer to utilize custom hooks for filtering and grouping countries
- Replaced inline filtering and grouping logic in CountryDrawer with useCountryFiltering and useCountryGrouping hooks for improved modularity and readability.
- Introduced useCountryFiltering to handle country search filtering.
- Added useCountryGrouping to manage country grouping by the first letter, enhancing organization of country options.
