+++ packages/mini-apps-ui-kit-react/src/components/CountryDrawer/CountryDrawer.tsx
@@ -1,18 +1,14 @@
 import { useRef, useState } from "react";
-import { defaultCountries, parseCountry } from "react-international-phone";
 
 import { Button } from "../Button";
 import { Drawer, DrawerClose, DrawerContent, DrawerTrigger } from "../Drawer";
 import { CountryCode } from "../Flag";
 import { XMark } from "../Icons/XMark";
 import { GroupedCountryList } from "../PhoneField/GroupedCountryList";
-import {
-  filterCountries,
-  getCountryDataListByCodes,
-  getValidatedCountryCode,
-} from "../PhoneField/utils";
 import { SearchField } from "../SearchField";
 import { TopBar } from "../TopBar";
+import { useCountryFiltering } from "./useCountryFiltering";
+import { useCountryGrouping } from "./useCountryGrouping";
 
 interface CountryDrawerProps {
   value: CountryCode;
@@ -22,13 +18,6 @@ interface CountryDrawerProps {
   children: React.ReactNode;
   onChange: (countryCode: string) => void;
   onAnimationEnd?: (open: boolean) => void;
-}
-
-interface GroupedCountries {
-  [key: string]: {
-    countryCode: CountryCode;
-    name: string;
-  }[];
 }
 
 export function CountryDrawer({
@@ -48,25 +37,15 @@ export function CountryDrawer({
     setSearchText(event.target.value);
   };
 
-  const richCountries = countries ? getCountryDataListByCodes(countries) : defaultCountries;
+  const filteredCountries = useCountryFiltering({
+    countries,
+    searchText,
+  });
 
-  const filteredCountries = filterCountries(richCountries, searchText);
-
-  const groupedCountries = filteredCountries.reduce<GroupedCountries>((acc, country) => {
-    const parsedCountry = parseCountry(country);
-    const firstLetter = parsedCountry.name.charAt(0).toUpperCase();
-
-    if (!acc[firstLetter]) {
-      acc[firstLetter] = [];
-    }
-
-    acc[firstLetter].push({
-      countryCode: getValidatedCountryCode(parsedCountry.iso2, defaultValue),
-      name: parsedCountry.name,
-    });
-
-    return acc;
-  }, {});
+  const groupedCountries = useCountryGrouping({
+    countries: filteredCountries,
+    defaultValue,
+  });
 
   const handleCountrySelect = (countryCode: CountryCode) => {
     onChange(countryCode);
